+++ projects/gauntlet/index.js
@@ -1,4 +1,4 @@
-const { getCuratorExport } = require("../helper/curators");
+const { getCuratorExport, kaminoLendVaultTvl } = require("../helper/curators");
 const axios = require('axios');
 
 const configs = {
@@ -128,10 +128,9 @@ const configs = {
 
 // --- Drift Solana TVL logic ---
 const ADDRESSES = require('../helper/coreAssets.json')
-const { getMultipleAccounts, getProvider, getConnection, getTokenAccountBalances } = require('../helper/solana')
+const { getMultipleAccounts, getProvider, } = require('../helper/solana')
 const { Program, BN } = require("@project-serum/anchor")
 const { PublicKey } = require("@solana/web3.js")
-const { bs58 } = require('@project-serum/anchor/dist/cjs/utils/bytes')
 
 const TOKEN_INFO = {
   USDC: {
@@ -229,55 +228,8 @@ const VAULT_USER_ACCOUNTS = [
 ]
 
 // --- Kamino Lend Vault Layer ---
-const KAMINO_LEND_VAULT_LAYER_PROGRAM_ID = new PublicKey('KvauGMspG5k6rtzrqqn7WNn3oZdyKqLKwK2XWQ8FLjd')
 const GAUNTLET_ADMIN = new PublicKey('JC8sPweHaHr1kWzAvykaAmLsWtSWhi3M4NnyYGRdxgkt')
 
-async function kaminoLendVaultTvl(api) {
-  const connection = getConnection()
-  const provider = getProvider()
-
-  // Load the kvault IDL
-  const kvaultIdl = require('./kvault-idl.json')
-  const kvaultProgram = new Program(kvaultIdl, KAMINO_LEND_VAULT_LAYER_PROGRAM_ID, provider)
-
-  // Query vault accounts directly using getProgramAccounts with base58 encoded filter
-  const adminBytes = GAUNTLET_ADMIN.toBuffer()
-  const rawAccounts = await connection.getProgramAccounts(
-    KAMINO_LEND_VAULT_LAYER_PROGRAM_ID,
-    {
-      filters: [
-        {
-          memcmp: {
-            offset: 8, // After 8-byte Anchor discriminator
-            bytes: bs58.encode(adminBytes)
-          }
-        }
-      ]
-    }
-  )
-
-  // Extract vault data and calculate total AUM from prevAumSf
-  for (const { account } of rawAccounts) {
-    // Decode the account using Anchor's coder
-    const vaultState = kvaultProgram.coder.accounts.decode('VaultState', account.data)
-    const tokenMint = vaultState.tokenMint.toString()
-
-    // prevAumSf is in scaled fixed point format (60-bit fractional part)
-    // It represents the AUM in token's native units (already includes decimals)
-    // Total AUM = prevAumSf / (2^60)
-    const prevAumSf = vaultState.prevAumSf
-    const SCALING_FACTOR = BigInt(2) ** BigInt(60) // 2^60 for 60-bit fractional part
-
-    // Convert prevAumSf to BigInt (it's a u128, which Anchor returns as BN or string)
-    const aumSfBigInt = typeof prevAumSf === 'string' ? BigInt(prevAumSf) : BigInt(prevAumSf.toString())
-
-    // Calculate total tokens: prevAumSf / 2^60
-    // prevAumSf is already in token's smallest unit (with decimals), so we just divide by the scaling factor
-    const totalTokens = aumSfBigInt / SCALING_FACTOR
-
-    api.add(tokenMint, totalTokens.toString())
-  }
-}
 
 async function tvl(api) {
   const accounts = await getMultipleAccounts(VAULT_USER_ACCOUNTS)
@@ -307,7 +259,7 @@ async function tvl(api) {
   }
 
   // Kamino Lend vaults
-  await kaminoLendVaultTvl(api)
+  await kaminoLendVaultTvl(api, GAUNTLET_ADMIN)
 }
 
 async function megavaultTvl(api) {
