+++ projects/sentora/index.js
@@ -8,6 +8,12 @@ const configs = {
         '0x5aB5FE7d04CFDeFb9daf61f6f569a58A53D05eE1',
         '0xe78C246ea973389F55BAEADF71e04750D50417d1'
       ],
+      morphoVaultOwners: [
+        '0x13DE0cEE0B83562CBfD46682e10FfA4E3c5090e1',
+      ],
+    },
+    solana: {
+      kaminoLendVaultAdmins: ['7fLxEftpppneavpueYgP2s7HhSGbWpj2jTCmAEwwqonY'],
     },
   }
 }
